trying to master manual lax.conv_general_dilated_patches

+++ src/networks/perceiver.py
@@ -55,7 +55,7 @@ def scaled_dot_product(query: Array,
                        ) -> Array:
     attn = jnp.einsum('...qhd,...khd->...hqk', query, key)
     attn /= np.sqrt(query.shape[-1])
-    attn = jax.nn.softmax(attn)
+    attn = jax.nn.softmax(attn.astype(jnp.float32)).astype(value.dtype)
     return jnp.einsum('...hqk,...khd->...qhd', attn, value)
